insert, delete, and get api endpoints and db handlers are working, created errors, language codes, models, initial database migration files

+++ backend/db/book.go
@@ -15,16 +15,16 @@ func (c *DBController) InsertBook(title string, author string, pubDate time.Time
 
 // Deletes a book in the database
 // Takes the book id of the book to delete.
-func (c *DBController) DeleteBook(uid int) {}
+func (c *DBController) DeleteBook(uid int64) {}
 
 // Gets the book with id in the database
 // Returns the book object
-func (c *DBController) GetBook(bid int) *models.Book {
+func (c *DBController) GetBook(bid int64) *models.Book {
 	return nil
 }
 
 // Get the books that were exchanged by the user
 // Returns the book list the user
-func (c *DBController) GetUserBooks(uid int) []*models.Book {
+func (c *DBController) GetUserBooks(uid int64) []*models.Book {
 	return []*models.Book{}
 }
